Intermdeiate state.
