+++ routerd/src/zebra/linux/mod.rs
@@ -3,6 +3,7 @@
 //   Copyright (C) 2018,2019 Toshiaki Takada
 //
 
+pub mod rt_netlink;
 pub mod netlink;
 //pub mod address;
 //pub mod route;
